requirements.txt
# Python dependencies for SSF Transmitter Web Application
Flask==3.0.2
flask-cors==4.0.0
PyJWT==2.8.0
cryptography==42.0.5
requests==2.31.0
gunicorn==21.2.0

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
